Updated Google Sign-In Method
The previous API and code used will be deprecated in March 2023, therefore, with longevity in mind, I transitioned to Google Identity. Google Identity is their newest way to access Google Sign-In and authorize users. So far, you can ONLY move past the login page if you log in with a @nau.edu email. Any other gmail will not allow the user past the login screen. Once we begin implementing tokens and calling data, we will be able to send Students to their associated home page and Administrators to their associated admin home page. As of right now, both sign-in buttons send you to the Student home page upon success.