Altered the Landing page and Added Two new pages
I'm altering the landing page (index.html) to prompt the user to click on a button submission to indicate whether they are logging in as a student or as an admin. There is no security check *as of this commit* further than requiring an @nau.edu email. This will allow us to properly redirect to the student home page or the admin home page based on the login page the user is using. Once we being implementing the TokenID and checking the authorized email associated with the gmail account, then we will be able to distinguish between a student and an admin with a single Google Sign-In button and one landing page/login page.