Altered the index.html login screen
Changed the index.html "Landing Page" back into the main login page. I styled it, and the carrot. I also moved the form buttons lower and ideally off-screen, with the intention of it being for developer use only. The studentlogin.html page and the adminlogin.html page will be kept and used for testing purposes for the time being as we still have to work out the token code for both students and admins. I'm hoping to test the Google Identity token code through the main index.html page, which means the other two login pages are obsolete, but I'm leaving them just in case.