Updated Index page, Admin page and styling
Updated the index.html code to include the JWT code for when the connections are ready to be made. Started updating the admin page with features such as the search bar and the add student table. Styled the index page and admin page. Added some JS to the admin.js file as well.